# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for StreamShort, a short-form video streaming platform.",
        "title": "StreamShort API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/episodes/{id}/like": {
            "get": {
                "description": "Get the like status and count for an episode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Get Like Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LikeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Like or unlike an episode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Like/Unlike Episode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Like Action",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LikeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LikeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/me/liked-episodes": {
            "get": {
                "description": "Get list of episodes liked by the user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Get Liked Episodes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LikedEpisodesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/series/{id}/rating": {
            "post": {
                "description": "Rate a series (1-5 stars).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Rate Series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rating",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RatingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.LikeRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "\"like\" or \"unlike\" (legacy format)",
                    "type": "string"
                },
                "like": {
                    "description": "true = like, false = unlike (new format)",
                    "type": "boolean"
                }
            }
        },
        "handlers.LikeResponse": {
            "type": "object",
            "properties": {
                "is_liked": {
                    "type": "boolean"
                },
                "like_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.LikedEpisodeItem": {
            "type": "object",
            "properties": {
                "episode_id": {
                    "type": "string"
                },
                "episode_number": {
                    "type": "integer"
                },
                "episode_title": {
                    "type": "string"
                },
                "like_count": {
                    "type": "integer"
                },
                "liked_at": {
                    "type": "string"
                },
                "series_id": {
                    "type": "string"
                },
                "series_title": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        },
        "handlers.LikedEpisodesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.LikedEpisodeItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.RatingRequest": {
            "type": "object",
            "properties": {
                "rating": {
                    "description": "1-5 stars",
                    "type": "integer"
                }
            }
        },
        "handlers.RatingResponse": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_ratings": {
                    "type": "integer"
                }
            }
        },
...
    }
}